Fix tests.
Running the tests on a port other than 80 used to fail du to the use of
window.location.hostname. I changed it to window.location.host so the
correct port is used. I had to also change a line in
src/sc/dialog.coffee so the port is not used for the request to
soundcloud.com.