remove 'geo_blocked' and 'no_connection' errors
they no longer exist in the playback library, plus they would never be triggered before anyway.

If a track is geoblocked the `SC.stream()` promise rejects, because the api returns a 403 at that point.

We also always retry to retrieve urls on a connection timeout, whilst the player is not killed, so the 'no_connection' error would never be triggered